✂️ #689 Remove GRPC Client and Config (#714)
* commit before branch change

* replacing GRPC to HTTP method for public single program, private single program and private program list endpoints

* remove comments

* add joinProgramInvite http fetch and update resolver. some resolved error remain. git commit before switching branch

* fix error from date values and waiting program service to fix first and last name

* remove unused key value pairs in joinProgramInvite response

* updated name

* edit http fetch to match new Program Service api data format

* remove testing console.log

* add userList resolver and http fetch

* create program endpoint

* createProgram client is completed

* update inviteUser to use HTTP

* commit before branch switch

* commit before chaning branch

* joinProgram, updateUser fetch and resolver working, updateProgram waiting for testing

* Add Copyright

* Fix Empty Field Errors

* Edit

* Working Regions

* Remove Whitespace

* Fix Name Values

* Make required User fields non-nullable

* Remove References to GRPC

* Consolidate programService index file

* Remove old GRPC Comments

* Data Centers / Update Program Syntax improvement

* Fix Copyright Date

---------
