+++ .env.schema
@@ -20,7 +20,6 @@ VAULT_ROLE=argo_gateway
 
 # Ego
 EGO_ROOT_REST=
-EGO_ROOT_GRPC=
 EGO_DACO_POLICY_NAME=
 EGO_VAULT_SECRET_PATH=
 EGO_CLIENT_ID=
