+++ package.json
@@ -27,7 +27,6 @@
 		"@babel/plugin-proposal-optional-chaining": "^7.18.6",
 		"@babel/register": "^7.18.6",
 		"@elastic/elasticsearch": "^7.17.0",
-		"@grpc/proto-loader": "^0.6.13",
 		"@icgc-argo/ego-token-utils": "^8.2.0",
 		"@icgc-argo/program-service-proto": "0.1.0",
 		"apollo-link-context": "^1.0.20",
@@ -48,7 +47,6 @@
 		"form-data": "^4.0.0",
 		"graphql": "^14.7.0",
 		"graphql-tools": "^4.0.4",
-		"grpc": "^1.24.11",
 		"http-proxy-middleware": "^1.3.1",
 		"json-stringify-deterministic": "^1.0.7",
 		"lodash": "^4.17.21",
